pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autolab"
version = "0.1.0"
description = "Control plane for AI-operated RL training"
authors = [{name = "Aritra Ghosh"}]
requires-python = ">=3.11"
dependencies = [
    # Core training stack
    "torch>=2.4.0",
    "transformers>=4.51.0",
    "trl>=0.16.0,<0.20.0",
    "datasets>=3.0.0",
    "accelerate>=1.0.0",
    "peft>=0.13.0",
    "bitsandbytes>=0.44.0",
    # Reward verification
    "math-verify>=0.9.0",
    "latex2sympy2-extended",
    # Anthropic API client
    "anthropic>=0.40.0",
    # Config / utilities
    "pyyaml>=6.0",
    "safetensors>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["autolab"]

# flash-attn must be installed separately (build-from-source):
#   pip install flash-attn --no-build-isolation
# Then set attn_implementation: "flash_attention_2" in the YAML config.
# Remove that key from model_init_kwargs if flash-attn is not installed.
